+++ src/Raven.CodeAnalysis/Syntax/InternalSyntax/Parser/Parsers/StatementSyntaxParser.cs
@@ -147,7 +147,7 @@ internal class StatementSyntaxParser : SyntaxParser
 
     private static bool IsMacroExpansionKeyword(SyntaxToken token)
         => token.IsKind(SyntaxKind.IdentifierToken) &&
-           token.GetValueText() is "expand" or "replace" or "introduce";
+           token.GetValueText() is "expand" or "replace" or "introduce" or "fragment";
 
     private MacroExpansionStatementSyntax ParseMacroExpansionStatementSyntax()
     {